// src/siteapi/assetdelivery.v2/models/RobloxWebAssetsBatchAssetRequestItem.ts
export class RobloxWebAssetsBatchAssetRequestItem {
    'assetName'?: string;
    'assetType'?: string;
    'clientInsert'?: boolean;
    'placeId'?: number;
    'requestId'?: string;
    'scriptInsert'?: boolean;
    'serverPlaceId'?: number;
    'universeId'?: number;
    'accept'?: string;
    'encoding'?: string;
    'hash'?: string;
    'userAssetId'?: number;
    'assetId'?: number;
    'version'?: number;
    'assetVersionId'?: number;
    'modulePlaceId'?: number;
    'assetFormat'?: string;
    'robloxAssetFormat'?: string;
    'contentRepresentationPriorityList'?: string;
    'doNotFallbackToBaselineRepresentation'?: boolean;

    static readonly discriminator: string | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "assetName",
            "baseName": "assetName",
            "type": "string",
            "format": ""
        },
        {
            "name": "assetType",
            "baseName": "assetType",
            "type": "string",
            "format": ""
        },
        {
            "name": "clientInsert",
            "baseName": "clientInsert",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "placeId",
            "baseName": "placeId",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "requestId",
            "baseName": "requestId",
            "type": "string",
            "format": ""
        },
        {
            "name": "scriptInsert",
            "baseName": "scriptInsert",
            "type": "boolean",
            "format": ""
        },
        {
            "name": "serverPlaceId",
            "baseName": "serverPlaceId",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "universeId",
            "baseName": "universeId",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "accept",
            "baseName": "accept",
            "type": "string",
            "format": ""
        },
        {
            "name": "encoding",
            "baseName": "encoding",
            "type": "string",
            "format": ""
        },
        {
            "name": "hash",
            "baseName": "hash",
            "type": "string",
            "format": ""
        },
        {
            "name": "userAssetId",
            "baseName": "userAssetId",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "assetId",
            "baseName": "assetId",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "version",
            "baseName": "version",
            "type": "number",
            "format": "int32"
        },
        {
            "name": "assetVersionId",
            "baseName": "assetVersionId",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "modulePlaceId",
            "baseName": "modulePlaceId",
            "type": "number",
            "format": "int64"
        },
        {
            "name": "assetFormat",
            "baseName": "assetFormat",
            "type": "string",
            "format": ""
        },
        {
            "name": "robloxAssetFormat",
            "baseName": "roblox-assetFormat",
            "type": "string",
            "format": ""
        },
        {
            "name": "contentRepresentationPriorityList",
            "baseName": "contentRepresentationPriorityList",
            "type": "string",
            "format": ""
        },
        {
            "name": "doNotFallbackToBaselineRepresentation",
            "baseName": "doNotFallbackToBaselineRepresentation",
            "type": "boolean",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return RobloxWebAssetsBatchAssetRequestItem.attributeTypeMap;
    }

    public constructor() {
    }
}